+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ClassWarningSuppressor.java
@@ -2,6 +2,8 @@ package edu.umd.cs.findbugs;
 
 public class ClassWarningSuppressor extends WarningSuppressor {
 
+    private final static String BUG_TYPE = "US_USELESS_SUPPRESSION_ON_CLASS";
+
     ClassAnnotation clazz;
 
     public ClassWarningSuppressor(String bugPattern, ClassAnnotation clazz) {
@@ -14,6 +16,12 @@ public class ClassWarningSuppressor extends WarningSuppressor {
 
     public ClassAnnotation getClassAnnotation() {
         return clazz;
+    }
+
+    @Override
+    public BugInstance buildUselessSuppressionBugInstance() {
+        return new BugInstance(BUG_TYPE, PRIORITY)
+                .addClass(clazz.getClassDescriptor());
     }
 
     @Override