+++ spotbugsTestCases/src/java/suppression/Issue641.java
@@ -1,0 +1,21 @@
+package suppression;
+
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+
+@SuppressFBWarnings(value = "UUF_UNUSED_FIELD") // <- useless suppression
+public class Issue641 {
+
+    @SuppressFBWarnings(value = "UUF_UNUSED_FIELD") // <- useless suppression
+    private String field;
+
+    @SuppressFBWarnings(value = "UUF_UNUSED_FIELD") // <- valid suppression
+    private String field2;
+
+    public String getField() {
+        return field;
+    }
+
+    public void setField(/* useless suppression -> */ @SuppressFBWarnings(value = "UUF_UNUSED_FIELD") String val) {
+        field = val;
+    }
+}